+++ e2e/cypress/integration/system_console/main_menu_spec.js
@@ -81,5 +81,14 @@ describe('Main menu', () => {
         // * Verify build date exists in modal
         cy.findByText('Build Date:').should('be.visible');
     });
+
+    it('MM-T899 - Edition and License: Verify Privacy Policy link points to correct URL', () => {
+        goToAdminConsole();
+
+        // * Find privacy link and then assert that the href is what we expect it to be
+        cy.findByTestId('privacyPolicyLink').then((el) => {
+            expect(el[0].href).equal('https://about.mattermost.com/default-privacy-policy/');
+        });
+    });
 });
