+++ e2e/cypress/integration/system_console/reporting/site_statistics_spec.js
@@ -11,6 +11,12 @@
 
 import * as TIMEOUTS from '../../../fixtures/timeouts';
 
+// # Goes to the System Scheme page as System Admin
+const goToAdminConsole = () => {
+    cy.apiAdminLogin();
+    cy.visit('/admin_console');
+};
+
 describe('System Console > Site Statistics', () => {
     it('MM-T904_1 Site Statistics displays expected content categories', () => {
         // # Require license.
@@ -69,6 +75,133 @@ describe('System Console > Site Statistics', () => {
         // * Check that the values for the stats are valid.
         cy.get('.admin-console__content .row').eq(0).find('.content').each((el) => {
             cy.wrap(el).eq(0).invoke('text').then(parseFloat).should('be.gt', 0);
+        });
+    });
+
+    it('MM-T902 - Reporting ➜ Site statistics line graphs show same date', () => {
+        // # Require license.
+        cy.apiRequireLicense();
+
+        goToAdminConsole();
+
+        // * Find site statistics and click it
+        cy.findByTestId('reporting.system_analytics').click();
+
+        let totalPostsDataSet;
+        let totalPostsFromBots;
+        let activeUsersWithPosts;
+
+        // # Grab all data from the 3 charts from there data labels
+        cy.findByTestId('totalPosts').then((el) => {
+            totalPostsDataSet = el[0].dataset.labels;
+            cy.findByTestId('totalPostsFromBots').then((el2) => {
+                totalPostsFromBots = el2[0].dataset.labels;
+                cy.findByTestId('activeUsersWithPosts').then((el3) => {
+                    activeUsersWithPosts = el3[0].dataset.labels;
+
+                    // * Assert that all the dates are the same
+                    expect(totalPostsDataSet).equal(totalPostsFromBots);
+                    expect(totalPostsDataSet).equal(activeUsersWithPosts);
+                    expect(totalPostsFromBots).equal(activeUsersWithPosts);
+                });
+            });
+        });
+    });
+
+    it('MM-T903 - Site Statistics > Deactivating a user increments the Daily and Monthly Active Users counts down', () => {
+        cy.apiInitSetup().then(({team, user}) => {
+            const testUser = user;
+            const testTeam = team;
+
+            // # Login as test user and visit town-square
+            cy.apiLogin(testUser);
+            cy.visit(`/${testTeam.name}/channels/town-square`);
+
+            // # Wait two seconds then go to admin console
+            cy.wait(TIMEOUTS.TWO_SEC);
+            goToAdminConsole();
+
+            // # Go to system analytics
+            cy.get('#reporting\\/system_analytics').click();
+            cy.wait(TIMEOUTS.ONE_SEC);
+
+            let totalActiveUsersInitial;
+            let dailyActiveUsersInital;
+            let monthlyActiveUsersInital;
+            let totalActiveUsersFinal;
+            let dailyActiveUsersFinal;
+            let monthlyActiveUsersFinal;
+
+            // # Get the number and turn them into numbers
+            cy.findByTestId('totalActiveUsers').invoke('text').then((text) => {
+                totalActiveUsersInitial = parseInt(text, 10);
+                cy.findByTestId('dailyActiveUsers').invoke('text').then((text2) => {
+                    dailyActiveUsersInital = parseInt(text2, 10);
+                    cy.findByTestId('monthlyActiveUsers').invoke('text').then((text3) => {
+                        monthlyActiveUsersInital = parseInt(text3, 10);
+
+                        // # Deactivate user and relaod page and then wait 2 seconds
+                        cy.externalActivateUser(testUser.id, false);
+                        cy.reload();
+                        cy.wait(TIMEOUTS.TWO_SEC);
+
+                        // # Get the numbers required again
+                        cy.findByTestId('totalActiveUsers').invoke('text').then((text4) => {
+                            totalActiveUsersFinal = parseInt(text4, 10);
+
+                            cy.findByTestId('dailyActiveUsers').invoke('text').then((text5) => {
+                                dailyActiveUsersFinal = parseInt(text5, 10);
+
+                                cy.findByTestId('monthlyActiveUsers').invoke('text').then((text6) => {
+                                    monthlyActiveUsersFinal = parseInt(text6, 10);
+
+                                    // * Assert that the final number is the initial number minus one
+                                    expect(totalActiveUsersFinal).equal(totalActiveUsersInitial - 1);
+                                    expect(dailyActiveUsersFinal).equal(dailyActiveUsersInital - 1);
+                                    expect(monthlyActiveUsersFinal).equal(monthlyActiveUsersInital - 1);
+                                });
+                            });
+                        });
+                    });
+                });
+            });
+        });
+    });
+
+    it('MM-T905 - Site Statistics card labels in different languages', () => {
+        cy.apiInitSetup().then(({team}) => {
+            const testTeam = team;
+
+            // # Login as admin and set the langauge to french
+            cy.apiAdminLogin();
+            cy.visit(`/${testTeam.name}/channels/town-square`);
+            cy.get('#headerUsername').click();
+            cy.get('#accountSettings').should('be.visible').click();
+            cy.get('#displayButton').click();
+            cy.get('#languagesEdit').click();
+            cy.get('#displayLanguage').type('Français{enter}');
+            cy.get('#saveSetting').click();
+
+            // * Once in site statistics, check and make sure the boxes are truncated or not according to image on test
+            cy.visit('/admin_console/reporting/system_analytics');
+
+            const testIds = ['totalActiveUsersTitle', 'totalTeamsTitle', 'totalChannelsTitle', 'totalPostsTitle', 'totalSessionsTitle', 'totalCommandsTitle', 'incomingWebhooksTitle',
+                'outgoingWebhooksTitle', 'dailyActiveUsersTitle', 'monthlyActiveUsersTitle', 'websocketConnsTitle', 'masterDbConnsTitle', 'replicaDbConnsTitle'];
+
+            testIds.forEach((id) => {
+                let expectedResult = false;
+                if (id === 'totalCommandsTitle' || id === 'masterDbConnsTitle' || id === 'replicaDbConnsTitle') {
+                    expectedResult = true;
+                }
+
+                cy.findByTestId(id).then((el) => {
+                    const titleSpan = el[0].childNodes[0];
+
+                    // * All the boxes on System Statistics page should have UNTRUNCATED titles when in french except Total Commands, Master DB Conns, and Replica DB Conns.
+                    // * The following asserts if the they are truncated or not. If false, it means they are not truncated. If true, they are truncated.
+                    expect(titleSpan.scrollWidth > titleSpan.clientWidth).equal(expectedResult);
+                });
+            });
         });
     });
 });
